% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_to_clipboard.R
\name{colnames_to_clipboard}
\alias{colnames_to_clipboard}
\title{get df colnames}
\usage{
colnames_to_clipboard(df = NULL)
}
\arguments{
\item{df}{The name of a \code{data.frame} or a \code{data.frame} itself, if not provided, an addin is launched.}
}
\description{
get df colnames
}
\examples{
\dontrun{

# Launch addin

# or pass the name of a data.frame
colnames_to_clipboard("iris")

# or directly the object
colnames_to_clipboard(iris)

}
}
