% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_cell.R
\name{color_cell}
\alias{color_cell}
\title{color the cell}
\usage{
color_cell(df, ft, col, fn = min)
}
\arguments{
\item{df}{dataframe}

\item{ft}{flextable object}

\item{col}{the coulumns names for formate}

\item{fn}{the fuction to locate, Default: min}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
color a specific location cell
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{pull}}
 \code{\link[flextable]{color}},\code{\link[flextable]{bold}}
}
