% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_min_max.R
\name{color_min_max}
\alias{color_min_max}
\title{color the cell with min and max values}
\usage{
color_min_max(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
flextable
}
\description{
color the cell with min and max values
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #iris \%>\%
    head() \%>\%
    color_min_max() \%>\%
    manuscriptsJX::my_ft()
 }
}
}
\seealso{
\code{\link[flextable]{flextable}}
 \code{\link[dplyr]{select_all}}
}
