% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_tableau}
\alias{color_tableau}
\title{color palates for ggplot2}
\usage{
color_tableau(palette = "Tableau 10", type = "regular", direction = 1, ...)
}
\description{
color palates for ggplot2
}
\examples{

}
