% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix_to_heatmap.R
\name{confusion_matrix_to_heatmap}
\alias{confusion_matrix_to_heatmap}
\title{turn matrix to heatmap}
\usage{
confusion_matrix_to_heatmap(confusion_mat, class)
}
\arguments{
\item{confusion_mat}{matrix from python machine learning}

\item{class}{vector, the y class order(must the right order)}
}
\value{
ggplot object
}
\description{
turn matrix to heatmap
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{rownames}}
 \code{\link[tidyr]{gather}}
 \code{\link[ggplot2]{ggplot}}
}
