% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_to_pct.R
\name{count_to_pct}
\alias{count_to_pct}
\title{add percent to count}
\usage{
count_to_pct(data, ..., col = n)
}
\arguments{
\item{data}{dataframe}
}
\value{
dataframe
}
\description{
add percent to count
}
\examples{

iris \%>\%
    count(Species) \%>\%
    count_to_pct()
}
