% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cross_paste0}
\alias{cross_paste0}
\title{Create a concatenation cross product of chracter vectors}
\usage{
cross_paste0(chars1, chars2)
}
\arguments{
\item{chars1}{A character vector of string prefixes}

\item{chars2}{A character vector of string suffixes}
}
\value{
A character vector of all prefix/suffix combos.
}
\description{
For every string in two character vectors create the full
outer product of pasting every string in `chars1`` before
every string in `chars2`. This version was created by
Jono Carrol.
}
\examples{
\dontrun{
manuscriptsJX::cross_paste0(c("jixing", "jiren", "jide"),
                           c("_hezhong", "_yihui"))
#> [1] "jixing_hezhong" "jixing_yihui"   "jiren_hezhong"  "jiren_yihui"
#> [5] "jide_hezhong"   "jide_yihui"
}

}
