% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_UTF8.R
\name{df_to_UTF8}
\alias{df_to_UTF8}
\title{Turn GBK or other char coding to UTF-8}
\usage{
df_to_UTF8(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame
}
\description{
Turn GBK or other char coding to UTF-8
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{set_names}}
 \code{\link[tmcn]{toUTF8}}
 \code{\link[dplyr]{mutate_all}}
}
