% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_turn.R
\name{df_turn}
\alias{df_turn}
\title{t() for data frame}
\usage{
df_turn(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame
}
\description{
t() for data frame
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
