% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_read_annotate_format.R
\name{excel_read_annotate_format}
\alias{excel_read_annotate_format}
\title{Annotate Meaningful Formatting}
\usage{
excel_read_annotate_format(xlfilepath, orig, new)
}
\arguments{
\item{xlfilepath}{path to spreadsheet file (xlx or xlsx)}

\item{orig}{Variable to annotate formatting in}

\item{new}{Name of new variable with cell formatting pasted as a string.}
}
\value{
dataframe
}
\description{
Annotate Meaningful Formatting
}
\examples{
example_spreadsheet <- system.file("extdata/dog_test.xlsx", package = "unheadr")
excel_read_annotate_format(example_spreadsheet,orig = Task, new=Task_annotated)


}
