% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_word_table.R
\name{extract_word_table}
\alias{extract_word_table}
\title{extract table from word}
\usage{
extract_word_table(doc_path)
}
\arguments{
\item{doc_path}{doc file papth}
}
\value{
table list
}
\description{
extrac word from word
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[officer]{read_docx}}
 \code{\link[tibble]{as_tibble}}
 \code{\link[dplyr]{filter}}
 \code{\link[tidyr]{pivot_wider}}
 \code{\link[purrr]{set_names}}
}
