% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numb_comma.R
\name{format_comma}
\alias{format_comma}
\title{turn number display with comma}
\usage{
format_comma(x, mark = ",", prefix = "", ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{mark}{PARAM_DESCRIPTION, Default: ','}

\item{prefix}{PARAM_DESCRIPTION, Default: ''}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
turn number display with comma
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
