% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_merge_colnames2header.R
\name{ft_merge_colnames2header}
\alias{ft_merge_colnames2header}
\title{colnames merge to header}
\usage{
ft_merge_colnames2header(df, sep = "\\\\.")
}
\arguments{
\item{df}{dataframe}

\item{sep}{the separator, Default: '\.'}
}
\value{
flextable object
}
\description{
This function is use to merge the dataframe colname the same element as head. first, i seprate the colnames with sep as dataframe typology. them I use flextable function to merge the same element.
}
\details{
DETAILS
}
\examples{
head(iris) \%>\%
ft_merge_colnames2header(., sep = "\\\\.")
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate}},\code{\link[dplyr]{if_else}}
 \code{\link[tidyr]{separate}}
 \code{\link[flextable]{flextable}},\code{\link[flextable]{set_header_footer_df}},\code{\link[flextable]{merge_h}},\code{\link[flextable]{merge_v}},\code{\link[flextable]{theme_booktabs}},\code{\link[flextable]{autofit}},\code{\link[flextable]{fix_border_issues}},\code{\link[flextable]{align}}
}
