% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df.R
\name{get_df}
\alias{get_df}
\title{Retrieve a data.frame by name from an environment}
\usage{
get_df(df, env = globalenv())
}
\arguments{
\item{df}{character, name of the object}

\item{env}{an environment, Default: globalenv()}
}
\value{
the object
}
\description{
Retrieve a data.frame by name from an environment
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
