% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_to_df.R
\name{gtf_to_df}
\alias{gtf_to_df}
\title{from gtf file to data frame}
\usage{
gtf_to_df(path_gtf)
}
\arguments{
\item{path_gtf}{path to gtf file}
}
\value{
data frame
}
\description{
from gtf file to data frame
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rtracklayer]{io}}
 \code{\link[tidyr]{separate}}
}
