% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_to_pdf.R
\name{images_to_pdf}
\alias{images_to_pdf}
\title{FUNCTION_TITLE}
\usage{
images_to_pdf(path_dir, pdf_name = "merge")
}
\arguments{
\item{path_dir}{PARAM_DESCRIPTION}

\item{pdf_name}{PARAM_DESCRIPTION, Default: 'merge'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[fs]{dir_ls}}
 \code{\link[stringr]{str_c}},\code{\link[stringr]{str_glue}}
}
