% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_info_extract.R
\name{meta_info_extract}
\alias{meta_info_extract}
\title{get dataframe meta infomation}
\usage{
meta_info_extract(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame
}
\description{
get dataframe meta infomation
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 meta_info_extract(iris)
 }
}
}
\seealso{
\code{\link[dplyr]{summarise_all}}
 \code{\link[tidyr]{gather}}
 \code{\link[labelled]{var_label}}
}
