% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_boxcox.R
\name{my_boxcox}
\alias{my_boxcox}
\title{transform variable to normal variable}
\usage{
my_boxcox(y)
}
\arguments{
\item{y}{vector}
}
\value{
vector
}
\description{
transform variable to normal variable
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[MASS]{boxcox}}
}
