% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_boxplot.R
\name{my_boxplot}
\alias{my_boxplot}
\title{FUNCTION_TITLE}
\usage{
my_boxplot(df, ...)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 # Box plots with jittered points
 }
}
}
\seealso{
\code{\link[purrr]{map}}
 \code{\link[manuscriptsR]{df_turn}}
 \code{\link[dplyr]{filter}}
 \code{\link[tidystringdist]{tidy_comb_all}}
 \code{\link[ggpubr]{ggboxplot}}
 \code{\link[ggsci]{scale_color_lancet}}
}
