% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_clean_factor.R
\name{my_clean_factor}
\alias{my_clean_factor}
\title{change factor variable code}
\usage{
my_clean_factor(var, levels_code_string = "0=否, 1=是", ...)
}
\arguments{
\item{var}{variable name}

\item{levels_code_string}{a string for code change, Default: '0=否, 1=是'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
change factor variable code
}
\details{
DETAILS
}
\examples{
ABData <- tribble(
    ~name, ~sex, ~seek,
    "Mary", "F", "M",
    "Henry", "M", "F"
) \%>\% as.data.frame()


df_repl <- tibble::tribble(
    ~var,         ~var_code,
    "sex",  "F=Woman, M=Man",
    "seek", "F=zhu, M=jixing"
)

for(i in 1:nrow(df_repl)) {
    print(df_repl$var[i])
    ABData <-
        ABData \%>\%
        mutate_at(df_repl$var[i],
                  my_clean_factor, levels = df_repl$var_code[i])
}

\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{set_names}}
 \code{\link[cleaner]{clean}}
}
