% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_combine_word.R
\name{my_combine_word}
\alias{my_combine_word}
\title{combine a list of word file}
\usage{
my_combine_word(files_docx, out)
}
\arguments{
\item{files_docx}{a list of word file}

\item{out}{the path of combined word file}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
combine a list of word file
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
