% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_DT.R
\name{my_DT}
\alias{my_DT}
\title{dataframe to DT}
\usage{
my_DT(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
dataframe to DT
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DT]{datatable}}
}
