% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_export.R
\name{my_export}
\alias{my_export}
\title{FUNCTION_TITLE}
\usage{
my_export(res, path = "~/Desktop")
}
\arguments{
\item{res}{PARAM_DESCRIPTION}

\item{path}{PARAM_DESCRIPTION, Default: '~/Desktop'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[manuscriptsR]{my_word}}
 \code{\link[stringr]{str_glue}}
 \code{\link[export]{graph2office}},\code{\link[export]{character(0)}}
 \code{\link[rio]{export}}
 \code{\link[readr]{read_lines}}
 \code{\link[fs]{create}}
}
