% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ft.R
\name{my_ft}
\alias{my_ft}
\title{formated the dataframe or flextable as my favorite}
\usage{
my_ft(df, col.keys = NULL, col.digits = NULL)
}
\arguments{
\item{df}{dataframe or flextable}

\item{col.keys}{PARAM_DESCRIPTION, Default: NULL}

\item{col.digits}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
formated the dataframe or flextable as my favorite
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[flextable]{flextable}},\code{\link[flextable]{colformat_num}},\code{\link[flextable]{border_outer}},\code{\link[flextable]{border_inner_h}},\code{\link[flextable]{border_inner_v}},\code{\link[flextable]{font}},\code{\link[flextable]{bold}},\code{\link[flextable]{align}},\code{\link[flextable]{autofit}}
 \code{\link[officer]{fp_border}}
}
