% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ggforest.R
\name{my_ggforest}
\alias{my_ggforest}
\title{FUNCTION_TITLE}
\usage{
my_ggforest(rt, forestCol = "#0D456B")
}
\arguments{
\item{rt}{PARAM_DESCRIPTION}

\item{forestCol}{PARAM_DESCRIPTION, Default: '#0D456B'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
