% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_tableone.R
\name{my_tableone}
\alias{my_tableone}
\title{create table one for publication}
\usage{
my_tableone(data_tidy, group)
}
\arguments{
\item{data_tidy}{dataframe}

\item{group}{group factor}
}
\value{
dataframe
}
\description{
create table one for publication
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_glue}}
 \code{\link[compareGroups]{compareGroups}},\code{\link[compareGroups]{createTable}}
}
