% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_word.R
\name{my_word}
\alias{my_word}
\title{Table to Word}
\usage{
my_word(
  df,
  para = "Just for test",
  docx_path = "~/Desktop/tmp.docx",
  open = TRUE
)
}
\arguments{
\item{df}{dataframe or flextable}

\item{para}{text to write, Default: 'Just for test'}

\item{docx_path}{path to write docx file, Default: '~/Desktop/tmp.docx'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Write a table to docx file
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[officer]{fp_text}},\code{\link[officer]{read_docx}},\code{\link[officer]{body_add_par}},\code{\link[officer]{body_add_fpar}},\code{\link[officer]{fpar}}
 \code{\link[flextable]{body_add_flextable}}
 \code{\link[manuscriptsR]{my_ft}}
 \code{\link[glue]{glue}}
}
