% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{na_na2x}
\alias{na_na2x}
\title{NA to something}
\usage{
na_na2x(s, x = 0)
}
\arguments{
\item{s}{The input vector (of arbitrary class)}

\item{x}{The elements to transform into NA resp. what to transform NA into, Default: 0}
}
\value{
A vector with same dimension and class as s
}
\description{
convert NA to something
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #na_na2x(na_x2na(c(1:10),5),5)
 }
}
}
