% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_numb_sum.R
\name{na_numb_sum}
\alias{na_numb_sum}
\title{sum of the na number in df}
\usage{
na_numb_sum(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
sum of the na number in df
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
