% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{na_plot_missing}
\alias{na_plot_missing}
\title{FUNCTION_TITLE}
\usage{
na_plot_missing(df)
}
\arguments{
\item{df}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[DataExplorer]{plot_missing}}
 \code{\link[ggplot2]{ggtheme}}
}
