% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_replace_in_df.R
\name{na_replace_in_df}
\alias{na_replace_in_df}
\title{na to x in df with all columns}
\usage{
na_replace_in_df(df, x)
}
\arguments{
\item{df}{dataframe}

\item{x}{the replace value}
}
\value{
dataframe
}
\description{
na to x in df with all columns
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate_all}}
 \code{\link[tidyr]{replace_na}}
}
