% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{na_x2na}
\alias{na_x2na}
\title{na_x2na}
\usage{
na_x2na(s, x = 0)
}
\arguments{
\item{s}{The input vector (of arbitrary class)}

\item{x}{The elements to transform into NA resp. what to transform NA into}
}
\value{
A vector with same dimension and class as s
}
\description{
convert vecter to NA
}
\examples{
na_x2na(1:10, 1:5)

}
