% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{na_x2na_ifLessX}
\alias{na_x2na_ifLessX}
\title{FUNCTION_TITLE}
\usage{
na_x2na_ifLessX(s, x = 0)
}
\arguments{
\item{s}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION, Default: 0}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
