% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA.R
\name{na_null2na}
\alias{na_null2na}
\title{NULL to NA}
\usage{
na_null2na(x)
}
\arguments{
\item{x}{list element}
}
\value{
list element
}
\description{
try to convert NULL to NA
}
\examples{
test <- list(1, 2, 3, NULL)
test \%>\% map(~ na_null2na(.x))

}
