% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_search.R
\name{obj_search}
\alias{obj_search}
\title{Search for object with specific class in an environment}
\usage{
obj_search(what = "data.frame", env = globalenv())
}
\arguments{
\item{what}{a class to look for, Default: 'data.frame'}

\item{env}{An environment, Default: globalenv()}
}
\value{
Character vector of the names of objects, NULL if none
}
\description{
Search for object with specific class in an environment
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #obj_search("data.frame")
 library(ggplot2)
  data("mpg")
  obj_search("data.frame")
 
 }
}
}
