% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj_size.R
\name{obj_size}
\alias{obj_size}
\title{calculate the size of objector}
\usage{
obj_size(x)
}
\arguments{
\item{x}{R object}
}
\value{
strings
}
\description{
calculate the size of objector
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
