% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickview.R
\docType{package}
\name{quickview}
\alias{manuscriptsR}
\alias{manuscriptsR-package}
\alias{quickview}
\title{Open datasets in a View tab}
\usage{
quickview()
}
\description{
This is a package for me to deal with some daily stuff.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jixing475/manuscriptsR}
  \item Report bugs at \url{https://github.com/jixing475/manuscriptsR/issues}
}

}
\author{
\strong{Maintainer}: Jixing Liu \email{jixing475@163.com}

}
\keyword{internal}
