% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bamboolib.R
\name{run_bamboolib}
\alias{run_bamboolib}
\title{FUNCTION_TITLE}
\usage{
run_bamboolib(data, dir = "~/Desktop/work")
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 #iris \%>\% manuscriptsR::run_bamboolib()
 }
}
}
\seealso{
\code{\link[feather]{read_feather}}
}
