% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{runCurrentRscript}
\alias{runCurrentRscript}
\title{runCurrentRscript}
\usage{
runCurrentRscript(
  path = path.expand(rstudioapi::getActiveDocumentContext()[["path"]]),
  command = "Rscript",
  outputFile = "output.txt",
  autoOpen = TRUE,
  addRhome = TRUE
)
}
\arguments{
\item{path}{\code{character(1)} string, specifying the path of the file to
be used as Rscript argument (ideally a path to an R script)}

\item{command}{\code{character(1)} to be executed, defaults to
\code{"Rscript"}, which is the originally intended use}

\item{outputFile}{\code{character(1)} string, specifying the name of the
file, into which the output produced by running the Rscript will be written}

\item{autoOpen}{\code{logical(1)} should we open the output automatically?}

\item{addRhome}{\code{logical(1)} add R home as prefix to command? Useful
if path to R is not added to the \code{"PATH"} environment variable}
}
\value{
side-effects
}
\description{
Wrapper around \code{\link{system2}} to run Rscript with
default arguments for easy use as an RStudio addin
}
\seealso{
\code{\link{system2}}
}
