% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sougou_weixin.R
\name{sougou_weixin}
\alias{sougou_weixin}
\title{download article from sougou weixin}
\usage{
sougou_weixin(searchkey, maxpage = 2)
}
\arguments{
\item{searchkey}{query string}

\item{maxpage}{PARAM_DESCRIPTION, Default: 2}
}
\value{
dataframe
}
\description{
download article from sougou weixin
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
