% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string2vector.R
\name{string2vector}
\alias{string2vector}
\title{turn string to vector}
\usage{
string2vector(string, sep)
}
\arguments{
\item{string}{PARAM_DESCRIPTION}

\item{sep}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
turn string to vector
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_split}}
}
