% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_split.R
\name{super_split}
\alias{super_split}
\title{super split}
\usage{
super_split(.data, ...)
}
\arguments{
\item{.data}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
split function plus from #' Lists are my secret weapon for reporting stats with knitr, https://www.tjmahr.com/lists-knitr-secret-weapon/. The first variable splits the list at depth 0, the second variable splits the sublists at depth 1 (which were created in the prior split), and so on.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{nse-defuse}},\code{\link[rlang]{as_name}}
 \code{\link[purrr]{map_if}}
}
