% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_to_doi.R
\name{title_to_doi}
\alias{title_to_doi}
\title{FUNCTION_TITLE}
\usage{
title_to_doi(
  dois = NULL,
  query = NULL,
  filter = NULL,
  offset = NULL,
  limit = NULL,
  sample = NULL,
  sort = NULL,
  order = NULL,
  facet = FALSE,
  cursor = NULL,
  cursor_max = 5000,
  .progress = "none",
  flq = NULL,
  select = NULL,
  async = FALSE,
  ...
)
}
\arguments{
\item{dois}{PARAM_DESCRIPTION, Default: NULL}

\item{query}{PARAM_DESCRIPTION, Default: NULL}

\item{filter}{PARAM_DESCRIPTION, Default: NULL}

\item{offset}{PARAM_DESCRIPTION, Default: NULL}

\item{limit}{PARAM_DESCRIPTION, Default: NULL}

\item{sample}{PARAM_DESCRIPTION, Default: NULL}

\item{sort}{PARAM_DESCRIPTION, Default: NULL}

\item{order}{PARAM_DESCRIPTION, Default: NULL}

\item{facet}{PARAM_DESCRIPTION, Default: FALSE}

\item{cursor}{PARAM_DESCRIPTION, Default: NULL}

\item{cursor_max}{PARAM_DESCRIPTION, Default: 5000}

\item{.progress}{PARAM_DESCRIPTION, Default: 'none'}

\item{flq}{PARAM_DESCRIPTION, Default: NULL}

\item{select}{PARAM_DESCRIPTION, Default: NULL}

\item{async}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tibble]{as_tibble}}
}
