% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un_separate_rows.R
\name{un_separate_rows}
\alias{un_separate_rows}
\title{factor to character}
\usage{
un_separate_rows(df, groupingVar, separator)
}
\description{
opposite to separate_rows() from tidyr
}
\examples{

library(magrittr)
data.frame(
    stringsAsFactors = FALSE,
    player = c("Allan Houston", NA, "Latrell Sprewell", NA, NA),
    teams = c(
        "Pistons", "Knicks", "Warriors", "Knicks",
        "Timberwolves"
    ),
    position = c("Shooting guard", NA, "Small forward", NA, NA)
) \%>\%
    manuscriptsJX::un_separate_rows(groupingVar = player, separator = ", ")
#> # A tibble: 2 x 3
#>   player           teams                          position
#>   <chr>            <chr>                          <chr>
#> 1 Allan Houston    Pistons, Knicks                Shooting guard
#> 2 Latrell Sprewell Warriors, Knicks, Timberwolves Small forward


}
