% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unheadr_embedded_subheaders.R
\name{unheadr_embedded_subheaders}
\alias{unheadr_embedded_subheaders}
\title{Rectangling  embedded subheaders}
\usage{
unheadr_embedded_subheaders(df, regex, orig, new)
}
\arguments{
\item{df}{a dataframe with embedded subheaders}

\item{regex}{Regular expression to match the subheaders}

\item{orig}{Variable containing the extraneous subheaders}

\item{new}{Name or variable that will contain the group values}
}
\value{
dataframe
}
\description{
Rectangling  embedded subheaders
}
\examples{
dat <- data.frame(
drink = c(
   "Cold Drinks", "Soda", "Water", "Juice", "Lemonade",
   "Hot Drinks", "Tea", "Coffee"
),
price = c(NA, 2.99, 1.99, 3.15, 2, NA, 3.99, 1.99), stringsAsFactors = FALSE
)
(dat, "Drinks$", drink, "beverage_type")

}
