% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unheadr_unbreak_rows.R
\name{unheadr_unbreak_rows}
\alias{unheadr_unbreak_rows}
\title{Merge break rows}
\usage{
unheadr_unbreak_rows(df, regex, ogcol, sep = " ")
}
\arguments{
\item{df}{A data frame with at least two contiguous rows to be merged}

\item{regex}{A regular expression to identify sets of rows to be merged, meant for the leading of the two contiguous rows.}

\item{ogcol}{Variable with the text strings to match.}

\item{sep}{Character string to separate the unified values}
}
\value{
dataframe
}
\description{
Merge break rows
}
\examples{
bball <-
    data.frame(
        stringsAsFactors = FALSE,
        v1 = c(
            "Player", NA, "Sleve McDichael", "Dean Wesrey",
            "Karl Dandleton"
        ),
        v2 = c("Most points", "in a game", "55", "43", "41"),
        v3 = c("Season", "(year ending)", "2001", "2000", "2010")
    )
unheadr_unbreak_rows(bball, "Most", v2)

}
