% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unheadr_unbreak_vlues.R
\name{unheadr_unbreak_vlues}
\alias{unheadr_unbreak_vlues}
\title{Unbreak values using regex to match the broken half of the value}
\usage{
unheadr_unbreak_vlues(
  df,
  regex,
  ogcol,
  newcol,
  sep = " ",
  .slice_groups = FALSE
)
}
\arguments{
\item{df}{A data frame with one or more values within a variable broken up across two rows.}

\item{regex}{Regular expression for matching the second half of the broken values.}

\item{ogcol}{Variable to unbreak.}

\item{newcol}{Name of the new variable with the unified values.}

\item{sep}{Character string to separate the unified values (default is space).}

\item{.slice_groups}{When `.slice_groups = FALSE` (the default), the extra rows and the variable with broken values will not be dropped.}
}
\value{
dataframe
}
\description{
Unbreak values using regex to match the broken half of the value
}
\examples{
data(primates2017_broken)
regex matches strings starting in lowercase (broken species epithets)
unheadr_unbreak_vlues(primates2017_broken, "^[a-z]", scientific_name, sciname_new)


}
