% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_detect_keep.R
\name{vec_detect_keep}
\alias{vec_detect_keep}
\title{keep element from vector by str_dectect}
\usage{
vec_detect_keep(vec, pattern, ...)
}
\arguments{
\item{vec}{vector}

\item{pattern}{the pattern to match}

\item{...}{PARAM_DESCRIPTION}
}
\value{
vector
}
\description{
keep element from vector by str_dectct
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_detect}}
}
