% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_detect_remove.R
\name{vec_detect_remove}
\alias{vec_detect_remove}
\title{delete element from vector by str_dectect}
\usage{
vec_detect_remove(vec, pattern, ...)
}
\arguments{
\item{vec}{vector}

\item{pattern}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
delete element from vector by str_dectct
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_detect}}
}
