% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2string.R
\name{vector2string}
\alias{vector2string}
\title{Convert vector to string}
\usage{
vector2string(vector, sep = ", ")
}
\arguments{
\item{vector}{PARAM_DESCRIPTION}

\item{sep}{PARAM_DESCRIPTION, Default: ', '}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Convert vector to string
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
